Revert buffer common CLI flag
